refactor: split runner executor diagnostics
