+++ crates/runner/src/executor/diagnostics.rs
@@ -1,908 +1,39 @@
 //! Runner executor diagnostics and post-job log preservation.
 //!
-//! This module owns best-effort diagnostics gathered around an agent process:
-//! environment key summaries, guest-authored error and failure diagnostic files,
-//! in-VM abnormal-exit resource probes, host and guest OOM checks, stdout stream
-//! truncation and overflow markers, and post-job guest log copies.
-//!
-//! Environment diagnostics are deliberately key-only. They may record counts,
-//! byte totals, sanitized key names, omitted-key counts, value lengths, and
-//! suspicious key names, but they must not log environment values. This keeps
-//! operator debugging signals separate from secret-bearing user, model, or
-//! connector environment data.
-//!
-//! Diagnostic collection should not mask the original run result. Failed reads,
-//! failed probes, and failed log copies are reported to operator logs when useful
-//! and otherwise remain best-effort. User-visible failure enrichment must flow
-//! through the existing structured failure fields such as `ExecutionFailure`
-//! and `ResourceFailureDiagnostics`.
-//!
-//! Real-time stdout stream logs and post-job guest log copies preserve different
-//! sources. The stream log is host-captured supervised process output; copied
-//! guest logs are guest-authored files preserved after the job.
+//! This facade keeps the executor-facing diagnostics API narrow while the
+//! implementation is split by diagnostic source and side-effect boundary.
 
-use std::collections::HashMap;
-use std::io::{self, SeekFrom};
-use std::path::{Path, PathBuf};
-use std::time::Duration;
+mod environment;
+mod exit;
+mod guest_files;
+mod guest_logs;
+mod oom;
+mod resource;
+mod stdout_stream;
 
-use guest_contracts::diagnostics::{
-    CliObservedExitDiagnostic, FAILURE_DIAGNOSTIC_SCHEMA_VERSION, FailureClass,
-    FailureDetailSource, FailureDiagnostic,
+pub(super) use environment::{build_agent_env_diagnostics, build_agent_env_key_diagnostics};
+pub(super) use exit::{
+    AgentBootstrapAbnormalExitLogContext, log_agent_abnormal_exit_env_diagnostics,
+    log_agent_bootstrap_abnormal_exit_diagnostics, log_agent_process_exit_summary,
+    should_collect_agent_abnormal_exit_diagnostics,
+    should_collect_unattributed_sigkill_resource_diagnostics,
+    should_log_agent_bootstrap_abnormal_exit_diagnostics,
 };
-use sandbox::{
-    CopyFileOptions, EXEC_OUTPUT_LIMIT_64_KIB, ExecRequest, ExecTermination, ProcessOutputReceiver,
-    Sandbox,
+pub(super) use guest_files::{
+    read_guest_cli_agent_session_id, read_guest_error_file, read_guest_failure_diagnostic_file,
 };
-use tokio::io::{AsyncReadExt, AsyncSeekExt, AsyncWriteExt};
-use tracing::{info, warn};
-
-use super::env::is_runner_owned_env_key;
-use super::session_id::{invalid_session_id_diagnostic_preview, is_valid_session_id};
-use super::session_restore::SessionRestoreDiagnostics;
-use super::{
-    AGENT_ABNORMAL_EXIT_DIAGNOSTIC_SCRIPT, AGENT_ABNORMAL_EXIT_DIAGNOSTIC_TIMEOUT,
-    AGENT_ENV_KEY_DIAGNOSTIC_LIMIT, AGENT_ENV_KEY_MAX_CHARS, BOOTSTRAP_SENSITIVE_ENV_KEYS,
-    DEFAULT_EXEC_TIMEOUT, EXIT_SIGKILL, EXIT_SIGNAL_KILL, GUEST_LOG_COPY_MAX_BYTES,
-    ResourceFailureDiagnostics, ResourceFailureKind, SMALL_GUEST_FILE_MAX_BYTES,
-    STDOUT_STREAM_LIMIT_MARKER, STDOUT_STREAM_OVERFLOW_MARKER, SandboxReuseResult,
-    guest_runtime_path,
+pub(super) use guest_logs::copy_guest_logs;
+#[cfg(test)]
+pub(super) use guest_logs::{GuestLogCopyFailureKind, guest_log_copy_failure_kind};
+#[cfg(test)]
+pub(super) use oom::host_dmesg_indicates_oom;
+pub(super) use oom::{check_host_oom, dmesg_indicates_oom};
+pub(super) use resource::collect_agent_abnormal_exit_diagnostics;
+#[cfg(test)]
+pub(super) use resource::parse_agent_abnormal_exit_resource_diagnostics;
+pub(super) use stdout_stream::{
+    AgentStdoutStreamDiagnostics, StdoutDrainReport,
+    append_stdout_stream_diagnostics_to_stream_log, drain_stdout_to_file,
 };
-use crate::helper_exec::{helper_exec_succeeded, helper_exec_termination_label};
-use crate::ids::RunId;
-use crate::paths::LogPaths;
-use crate::types::ExecutionContext;
-
-const AGENT_ENV_VALUE_SIZE_DIAGNOSTIC_LIMIT: usize = 5;
-
-#[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
-pub(super) struct AgentStdoutStreamDiagnostics {
-    pub(super) bytes_written: u64,
-    pub(super) chunk_truncated: bool,
-    pub(super) stream_overflowed: bool,
-}
-
-impl AgentStdoutStreamDiagnostics {
-    pub(super) fn is_empty(self) -> bool {
-        !self.chunk_truncated && !self.stream_overflowed
-    }
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub(super) struct AgentEnvDiagnostics {
-    pub(super) env_count: usize,
-    pub(super) env_bytes: usize,
-    pub(super) runner_owned_count: usize,
-    pub(super) external_count: usize,
-    pub(super) suspicious_keys: Vec<String>,
-    pub(super) largest_entries: Vec<AgentEnvValueSizeDiagnostics>,
-}
-
-impl AgentEnvDiagnostics {
-    pub(super) fn suspicious_keys_csv(&self) -> String {
-        self.suspicious_keys.join(",")
-    }
-
-    pub(super) fn largest_entries_csv(&self) -> String {
-        self.largest_entries
-            .iter()
-            .map(|entry| format!("{}:{}", entry.key, entry.value_len))
-            .collect::<Vec<_>>()
-            .join(",")
-    }
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub(super) struct AgentEnvValueSizeDiagnostics {
-    pub(super) key: String,
-    pub(super) value_len: usize,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub(super) struct AgentEnvKeyDiagnostics {
-    pub(super) logged_keys: Vec<String>,
-    pub(super) omitted_key_count: usize,
-}
-
-impl AgentEnvKeyDiagnostics {
-    pub(super) fn logged_keys_csv(&self) -> String {
-        self.logged_keys.join(",")
-    }
-}
-
-pub(super) fn build_agent_env_diagnostics(
-    env: &HashMap<String, String>,
-    user_env: &HashMap<String, String>,
-) -> AgentEnvDiagnostics {
-    let mut suspicious_keys: Vec<String> = BOOTSTRAP_SENSITIVE_ENV_KEYS
-        .iter()
-        .copied()
-        .filter(|key| user_env.contains_key(*key))
-        .map(sanitize_env_key_for_diagnostic)
-        .collect();
-    suspicious_keys.sort();
-
-    let runner_owned_count = env
-        .keys()
-        .filter(|key| is_runner_owned_env_key(key))
-        .count();
-    let env_bytes = env.iter().map(|(key, value)| key.len() + value.len()).sum();
-    let mut largest_entries: Vec<AgentEnvValueSizeDiagnostics> = env
-        .iter()
-        .map(|(key, value)| AgentEnvValueSizeDiagnostics {
-            key: sanitize_env_key_for_diagnostic(key),
-            value_len: value.len(),
-        })
-        .collect();
-    largest_entries.sort_by(|left, right| {
-        right
-            .value_len
-            .cmp(&left.value_len)
-            .then_with(|| left.key.cmp(&right.key))
-    });
-    largest_entries.truncate(AGENT_ENV_VALUE_SIZE_DIAGNOSTIC_LIMIT);
-
-    AgentEnvDiagnostics {
-        env_count: env.len(),
-        env_bytes,
-        runner_owned_count,
-        external_count: env.len().saturating_sub(runner_owned_count),
-        suspicious_keys,
-        largest_entries,
-    }
-}
-
-pub(super) fn build_agent_env_key_diagnostics(env: &[(String, String)]) -> AgentEnvKeyDiagnostics {
-    let mut keys: Vec<String> = env
-        .iter()
-        .map(|(key, _)| sanitize_env_key_for_diagnostic(key))
-        .collect();
-    keys.sort();
-
-    let logged_keys: Vec<String> = keys
-        .iter()
-        .take(AGENT_ENV_KEY_DIAGNOSTIC_LIMIT)
-        .cloned()
-        .collect();
-    let omitted_key_count = keys.len().saturating_sub(logged_keys.len());
-
-    AgentEnvKeyDiagnostics {
-        logged_keys,
-        omitted_key_count,
-    }
-}
-
-pub(super) fn sanitize_env_key_for_diagnostic(key: &str) -> String {
-    let mut chars = key.escape_debug();
-    let mut truncated = String::new();
-    for _ in 0..AGENT_ENV_KEY_MAX_CHARS {
-        let Some(ch) = chars.next() else {
-            return truncated;
-        };
-        truncated.push(ch);
-    }
-    if chars.next().is_some() {
-        truncated.push_str("...");
-    }
-    truncated
-}
-
-pub(super) fn should_collect_agent_abnormal_exit_diagnostics(
-    wait_cancelled: bool,
-    exit: &sandbox::ProcessExit,
-    stderr: &str,
-    failure_diagnostic: Option<&FailureDiagnostic>,
-    guest_error: Option<&str>,
-) -> bool {
-    !wait_cancelled
-        && process_failed(exit)
-        && exit.diagnostic.is_empty()
-        && stderr.is_empty()
-        && failure_diagnostic.is_none()
-        && guest_error.is_none()
-}
-
-pub(super) fn should_collect_unattributed_sigkill_resource_diagnostics(
-    wait_cancelled: bool,
-    exit: &sandbox::ProcessExit,
-    failure_diagnostic: Option<&FailureDiagnostic>,
-) -> bool {
-    let ExecTermination::Exited { exit_code } = exit.termination else {
-        return false;
-    };
-
-    !wait_cancelled
-        && exit_code != 0
-        && exit.diagnostic.is_empty()
-        && failure_diagnostic
-            .is_some_and(|diagnostic| unattributed_sigkill_cli_failure(diagnostic, exit_code))
-}
-
-fn unattributed_sigkill_cli_failure(diagnostic: &FailureDiagnostic, exit_code: i32) -> bool {
-    diagnostic.failure_class == FailureClass::CliNonzero
-        && diagnostic.cli_exit_code == Some(exit_code)
-        && diagnostic.cli_termination.is_none()
-        && diagnostic.failure_detail_source == Some(FailureDetailSource::FallbackExitCode)
-        && diagnostic
-            .cli_observed_exit
-            .as_ref()
-            .is_some_and(|observed_exit| {
-                observed_sigkill_matches_exit_code(observed_exit, exit_code)
-            })
-}
-
-fn observed_sigkill_matches_exit_code(
-    observed_exit: &CliObservedExitDiagnostic,
-    exit_code: i32,
-) -> bool {
-    observed_exit.is_sigkill()
-        && observed_exit.exit_code.is_none()
-        && observed_exit.signal_number == Some(EXIT_SIGNAL_KILL)
-        && observed_exit.mapped_exit_code == EXIT_SIGKILL
-        && exit_code == EXIT_SIGKILL
-}
-
-pub(super) fn should_log_agent_bootstrap_abnormal_exit_diagnostics(
-    wait_cancelled: bool,
-    exit: &sandbox::ProcessExit,
-    failure_diagnostic: Option<&FailureDiagnostic>,
-    guest_error: Option<&str>,
-) -> bool {
-    !wait_cancelled
-        && process_exited_nonzero(exit)
-        && failure_diagnostic.is_none()
-        && guest_error.is_none()
-}
-
-fn process_failed(exit: &sandbox::ProcessExit) -> bool {
-    !matches!(exit.termination, ExecTermination::Exited { exit_code: 0 })
-}
-
-fn process_exited_nonzero(exit: &sandbox::ProcessExit) -> bool {
-    matches!(exit.termination, ExecTermination::Exited { exit_code } if exit_code != 0)
-}
-
-fn process_exit_code(exit: &sandbox::ProcessExit) -> Option<i32> {
-    match exit.termination {
-        ExecTermination::Exited { exit_code } => Some(exit_code),
-        ExecTermination::TimedOut
-        | ExecTermination::Cancelled
-        | ExecTermination::StartFailed
-        | ExecTermination::WaitFailed => None,
-    }
-}
-
-const ROOTFS_FULL_AVAILABLE_KB_THRESHOLD: u64 = 1024;
-
-pub(super) fn parse_agent_abnormal_exit_resource_diagnostics(
-    stdout: &str,
-) -> Option<ResourceFailureDiagnostics> {
-    let mut diagnostics = ResourceFailureDiagnostics::default();
-
-    for line in stdout.lines() {
-        if let Some(filesystem) = parse_filesystem_usage_line(line) {
-            match filesystem.mount_point {
-                "/" => {
-                    diagnostics.guest_root_fs_used_percent = Some(filesystem.used_percent);
-                    diagnostics.guest_root_fs_available_kb = filesystem.available_kb;
-                }
-                "/home/user/workspace" => {
-                    diagnostics.guest_workspace_fs_used_percent = Some(filesystem.used_percent);
-                }
-                _ => {}
-            }
-        }
-
-        if diagnostics.guest_memory_available_mb.is_none() {
-            diagnostics.guest_memory_available_mb = parse_memory_available_mb(line);
-        }
-    }
-
-    if rootfs_is_clearly_full(&diagnostics) {
-        diagnostics.failure_kind = Some(ResourceFailureKind::GuestRootFilesystemFull);
-    }
-
-    (!diagnostics.is_empty()).then_some(diagnostics)
-}
-
-struct FilesystemUsage<'a> {
-    mount_point: &'a str,
-    used_percent: u16,
-    available_kb: Option<u64>,
-}
-
-fn parse_filesystem_usage_line(line: &str) -> Option<FilesystemUsage<'_>> {
-    let columns: Vec<&str> = line.split_whitespace().collect();
-    if columns.len() < 6 {
-        return None;
-    }
-
-    let mount_point = *columns.last()?;
-    if mount_point != "/" && mount_point != "/home/user/workspace" {
-        return None;
-    }
-
-    let used_percent = columns
-        .get(columns.len().saturating_sub(2))
-        .and_then(|value| parse_percent(value))?;
-    let available_kb = columns
-        .get(columns.len().saturating_sub(3))
-        .and_then(|value| parse_available_kb(value));
-
-    Some(FilesystemUsage {
-        mount_point,
-        used_percent,
-        available_kb,
-    })
-}
-
-fn parse_percent(value: &str) -> Option<u16> {
-    value.strip_suffix('%')?.parse().ok()
-}
-
-fn parse_available_kb(value: &str) -> Option<u64> {
-    if let Ok(kb) = value.parse() {
-        return Some(kb);
-    }
-
-    let suffix = value.chars().last()?;
-    let unit_multiplier = match suffix {
-        'K' | 'k' => 1,
-        'M' | 'm' => 1024,
-        'G' | 'g' => 1024 * 1024,
-        'T' | 't' => 1024 * 1024 * 1024,
-        _ => return None,
-    };
-    let numeric = &value[..value.len() - suffix.len_utf8()];
-    let amount: f64 = numeric.parse().ok()?;
-    if !amount.is_finite() || amount.is_sign_negative() {
-        return None;
-    }
-    Some((amount * unit_multiplier as f64).round() as u64)
-}
-
-fn parse_memory_available_mb(line: &str) -> Option<u64> {
-    let columns: Vec<&str> = line.split_whitespace().collect();
-    if columns.first() == Some(&"Mem:") && columns.len() >= 7 {
-        columns.get(6).and_then(|value| value.parse().ok())
-    } else if let ["MemAvailable:", kb, ..] = columns.as_slice() {
-        kb.parse::<u64>().ok().map(|value| value / 1024)
-    } else {
-        None
-    }
-}
-
-fn rootfs_is_clearly_full(diagnostics: &ResourceFailureDiagnostics) -> bool {
-    diagnostics
-        .guest_root_fs_used_percent
-        .is_some_and(|percent| percent >= 100)
-        || diagnostics
-            .guest_root_fs_available_kb
-            .is_some_and(|available_kb| available_kb <= ROOTFS_FULL_AVAILABLE_KB_THRESHOLD)
-}
-
-pub(super) fn log_agent_process_exit_summary(
-    run_id: RunId,
-    sandbox_id: &str,
-    reuse_result: SandboxReuseResult,
-    exit: &sandbox::ProcessExit,
-    env_diagnostics: &AgentEnvDiagnostics,
-    stdout_stream_diagnostics: AgentStdoutStreamDiagnostics,
-) {
-    info!(
-        run_id = %run_id,
-        sandbox_id = %sandbox_id,
-        sandbox_reuse_result = reuse_result.as_wire(),
-        termination = ?exit.termination,
-        exit_code = ?process_exit_code(exit),
-        stdout_len = exit.stdout.len(),
-        stderr_len = exit.stderr.len(),
-        stdout_truncated = exit.stdout_truncated,
-        stderr_truncated = exit.stderr_truncated,
-        stdout_stream_bytes = stdout_stream_diagnostics.bytes_written,
-        diagnostic_present = !exit.diagnostic.is_empty(),
-        stream_overflowed = exit.stream_overflowed,
-        env_count = env_diagnostics.env_count,
-        env_bytes = env_diagnostics.env_bytes,
-        largest_env_value_lengths = %env_diagnostics.largest_entries_csv(),
-        suspicious_env_keys = %env_diagnostics.suspicious_keys_csv(),
-        "agent process exit summary"
-    );
-}
-
-pub(super) fn log_agent_abnormal_exit_env_diagnostics(
-    run_id: RunId,
-    sandbox_id: &str,
-    reuse_result: SandboxReuseResult,
-    exit: &sandbox::ProcessExit,
-    env_diagnostics: &AgentEnvDiagnostics,
-    env_key_diagnostics: &AgentEnvKeyDiagnostics,
-) {
-    warn!(
-        run_id = %run_id,
-        sandbox_id = %sandbox_id,
-        sandbox_reuse_result = reuse_result.as_wire(),
-        termination = ?exit.termination,
-        exit_code = ?process_exit_code(exit),
-        env_count = env_diagnostics.env_count,
-        env_bytes = env_diagnostics.env_bytes,
-        runner_owned_env_count = env_diagnostics.runner_owned_count,
-        external_env_count = env_diagnostics.external_count,
-        largest_env_value_lengths = %env_diagnostics.largest_entries_csv(),
-        suspicious_env_keys = %env_diagnostics.suspicious_keys_csv(),
-        env_keys = %env_key_diagnostics.logged_keys_csv(),
-        omitted_env_key_count = env_key_diagnostics.omitted_key_count,
-        "agent abnormal exit env diagnostics"
-    );
-}
-
-pub(super) struct AgentBootstrapAbnormalExitLogContext<'a> {
-    pub(super) run_id: RunId,
-    pub(super) sandbox_id: &'a str,
-    pub(super) reuse_result: SandboxReuseResult,
-    pub(super) exit: &'a sandbox::ProcessExit,
-    pub(super) env_diagnostics: &'a AgentEnvDiagnostics,
-    pub(super) env_key_diagnostics: &'a AgentEnvKeyDiagnostics,
-    pub(super) stdout_stream_diagnostics: AgentStdoutStreamDiagnostics,
-    pub(super) session_restore_diagnostics: Option<&'a SessionRestoreDiagnostics>,
-}
-
-pub(super) fn log_agent_bootstrap_abnormal_exit_diagnostics(
-    context: AgentBootstrapAbnormalExitLogContext<'_>,
-) {
-    let resume_session_framework = context
-        .session_restore_diagnostics
-        .map(|diagnostics| diagnostics.framework)
-        .unwrap_or("none");
-    let resume_session_id = context
-        .session_restore_diagnostics
-        .map(|diagnostics| diagnostics.session_id.as_str())
-        .unwrap_or("");
-    let resume_session_bytes_in = context
-        .session_restore_diagnostics
-        .map(|diagnostics| diagnostics.bytes_in);
-
-    warn!(
-        run_id = %context.run_id,
-        sandbox_id = %context.sandbox_id,
-        sandbox_reuse_result = context.reuse_result.as_wire(),
-        termination = ?context.exit.termination,
-        exit_code = ?process_exit_code(context.exit),
-        stdout_len = context.exit.stdout.len(),
-        stderr_len = context.exit.stderr.len(),
-        stdout_truncated = context.exit.stdout_truncated,
-        stderr_truncated = context.exit.stderr_truncated,
-        stdout_stream_bytes = context.stdout_stream_diagnostics.bytes_written,
-        captured_stderr_present = !context.exit.stderr.is_empty(),
-        captured_stderr_truncated = context.exit.stderr_truncated,
-        diagnostic_present = !context.exit.diagnostic.is_empty(),
-        stream_overflowed = context.exit.stream_overflowed,
-        env_count = context.env_diagnostics.env_count,
-        env_bytes = context.env_diagnostics.env_bytes,
-        runner_owned_env_count = context.env_diagnostics.runner_owned_count,
-        external_env_count = context.env_diagnostics.external_count,
-        largest_env_value_lengths = %context.env_diagnostics.largest_entries_csv(),
-        suspicious_env_keys = %context.env_diagnostics.suspicious_keys_csv(),
-        env_keys = %context.env_key_diagnostics.logged_keys_csv(),
-        omitted_env_key_count = context.env_key_diagnostics.omitted_key_count,
-        resume_session_framework = %resume_session_framework,
-        resume_session_id = %resume_session_id,
-        resume_session_bytes_in = ?resume_session_bytes_in,
-        "agent bootstrap abnormal exit diagnostics"
-    );
-}
-
-pub(super) async fn collect_agent_abnormal_exit_diagnostics(
-    sandbox: &dyn Sandbox,
-    run_id: RunId,
-    sandbox_id: &str,
-    reuse_result: SandboxReuseResult,
-    exit_code: i32,
-) -> Option<ResourceFailureDiagnostics> {
-    let request = ExecRequest {
-        cmd: AGENT_ABNORMAL_EXIT_DIAGNOSTIC_SCRIPT,
-        timeout: AGENT_ABNORMAL_EXIT_DIAGNOSTIC_TIMEOUT,
-        env: &[],
-        sudo: true,
-        stdin_bytes: None,
-        output_limits: EXEC_OUTPUT_LIMIT_64_KIB,
-    };
-
-    match sandbox
-        .exec_with_diagnostic_label(&request, "agent-abnormal-exit-diagnostics")
-        .await
-    {
-        Ok(result) => {
-            let stdout = String::from_utf8_lossy(&result.stdout);
-            let stderr = String::from_utf8_lossy(&result.stderr);
-            let diagnostic_succeeded = helper_exec_succeeded(&result);
-            let resource_diagnostics = diagnostic_succeeded
-                .then(|| parse_agent_abnormal_exit_resource_diagnostics(&stdout))
-                .flatten();
-            let resource_failure_kind = resource_diagnostics
-                .and_then(|diagnostics| diagnostics.failure_kind)
-                .map(ResourceFailureKind::as_str);
-            let guest_root_fs_used_percent = resource_diagnostics
-                .and_then(|diagnostics| diagnostics.guest_root_fs_used_percent)
-                .map(u64::from);
-            let guest_root_fs_available_kb =
-                resource_diagnostics.and_then(|diagnostics| diagnostics.guest_root_fs_available_kb);
-            let guest_workspace_fs_used_percent = resource_diagnostics
-                .and_then(|diagnostics| diagnostics.guest_workspace_fs_used_percent)
-                .map(u64::from);
-            let guest_memory_available_mb =
-                resource_diagnostics.and_then(|diagnostics| diagnostics.guest_memory_available_mb);
-            warn!(
-                run_id = %run_id,
-                sandbox_id = %sandbox_id,
-                sandbox_reuse_result = reuse_result.as_wire(),
-                exit_code,
-                diagnostic_termination = helper_exec_termination_label(&result),
-                diagnostic_succeeded,
-                diagnostic_stdout_len = result.stdout.len(),
-                diagnostic_stderr_len = result.stderr.len(),
-                diagnostic_stdout_truncated = result.stdout_truncated,
-                diagnostic_stderr_truncated = result.stderr_truncated,
-                resource_failure_kind,
-                guest_root_fs_used_percent,
-                guest_root_fs_available_kb,
-                guest_workspace_fs_used_percent,
-                guest_memory_available_mb,
-                diagnostic_stdout = %stdout,
-                diagnostic_stderr = %stderr,
-                "agent abnormal exit in-vm diagnostics"
-            );
-            resource_diagnostics
-        }
-        Err(error) => {
-            warn!(
-                run_id = %run_id,
-                sandbox_id = %sandbox_id,
-                sandbox_reuse_result = reuse_result.as_wire(),
-                exit_code,
-                error = %error,
-                "failed to collect agent abnormal exit in-vm diagnostics"
-            );
-            None
-        }
-    }
-}
-pub(super) async fn read_guest_error_file(sandbox: &dyn Sandbox, run_id: RunId) -> Option<String> {
-    let error_path = match guest_runtime_path(
-        run_id,
-        guest_contracts::runtime_paths::checkpoint_error_file,
-    ) {
-        Ok(path) => path,
-        Err(e) => {
-            warn!(run_id = %run_id, error = %e, "failed to resolve guest error file path");
-            return None;
-        }
-    };
-    match sandbox
-        .read_file(&error_path, SMALL_GUEST_FILE_MAX_BYTES)
-        .await
-    {
-        Ok(Some(bytes)) if !bytes.is_empty() => {
-            let msg = String::from_utf8_lossy(&bytes).trim().to_string();
-            Some(msg).filter(|s| !s.is_empty())
-        }
-        _ => None,
-    }
-}
-
-/// Read structured guest failure diagnostics from the guest filesystem.
-///
-/// Diagnostics are optional and best-effort. They must never change the
-/// user-visible completion error or mask the original exit status.
-pub(super) async fn read_guest_failure_diagnostic_file(
-    sandbox: &dyn Sandbox,
-    run_id: RunId,
-) -> Option<FailureDiagnostic> {
-    let path = match guest_runtime_path(
-        run_id,
-        guest_contracts::runtime_paths::failure_diagnostic_file,
-    ) {
-        Ok(path) => path,
-        Err(e) => {
-            warn!(run_id = %run_id, error = %e, "failed to resolve guest failure diagnostic path");
-            return None;
-        }
-    };
-    match sandbox.read_file(&path, SMALL_GUEST_FILE_MAX_BYTES).await {
-        Ok(Some(bytes)) if !bytes.iter().all(|byte| byte.is_ascii_whitespace()) => {
-            match serde_json::from_slice::<FailureDiagnostic>(&bytes) {
-                Ok(diagnostic)
-                    if diagnostic.schema_version == FAILURE_DIAGNOSTIC_SCHEMA_VERSION =>
-                {
-                    Some(diagnostic)
-                }
-                Ok(diagnostic) => {
-                    warn!(
-                        run_id = %run_id,
-                        schema_version = diagnostic.schema_version,
-                        "ignoring guest failure diagnostic with unsupported schema version"
-                    );
-                    None
-                }
-                Err(e) => {
-                    warn!(run_id = %run_id, error = %e, "failed to parse guest failure diagnostic");
-                    None
-                }
-            }
-        }
-        Ok(_) => None,
-        Err(e) => {
-            warn!(run_id = %run_id, error = %e, "failed to read guest failure diagnostic");
-            None
-        }
-    }
-}
-
-/// Read the CLI-generated session ID from the guest filesystem.
-///
-/// The guest-agent writes the session ID to the guest runtime directory
-/// after the CLI emits its `system/init` event. On first runs (no
-/// `resume_session`), the runner uses this to park the VM for keep-alive.
-pub(super) async fn read_guest_cli_agent_session_id(
-    sandbox: &dyn Sandbox,
-    run_id: RunId,
-) -> Option<String> {
-    let path = match guest_runtime_path(run_id, guest_contracts::runtime_paths::session_id_file) {
-        Ok(path) => path,
-        Err(e) => {
-            warn!(run_id = %run_id, error = %e, "failed to resolve guest session id path");
-            return None;
-        }
-    };
-    match sandbox.read_file(&path, SMALL_GUEST_FILE_MAX_BYTES).await {
-        Ok(Some(bytes)) if !bytes.is_empty() => {
-            let id = String::from_utf8_lossy(&bytes).trim().to_string();
-            if id.is_empty() {
-                return None;
-            }
-            if !is_valid_session_id(&id) {
-                warn!(
-                    run_id = %run_id,
-                    session_id = %invalid_session_id_diagnostic_preview(&id),
-                    "ignoring invalid guest session ID"
-                );
-                return None;
-            }
-            Some(id)
-        }
-        _ => None,
-    }
-}
-
-/// Returns true if dmesg output indicates an OOM kill.
-pub(super) fn dmesg_indicates_oom(stdout: &str) -> bool {
-    let lower = stdout.to_lowercase();
-    lower.contains("out of memory") || lower.contains("oom-kill") || lower.contains("oom_reaper")
-}
-
-/// Check host dmesg for a cgroup OOM kill of a specific firecracker process.
-/// Reads the entire ring buffer (~512KB) directly — no shell wrapper needed
-/// since the pure function handles filtering.  Times out after 5s to avoid
-/// blocking if sudo hangs.
-pub(super) async fn check_host_oom(pid: u32) -> bool {
-    let result = tokio::time::timeout(Duration::from_secs(5), async {
-        tokio::process::Command::new("dmesg").output().await
-    })
-    .await;
-    match result {
-        Ok(Ok(out)) if out.status.success() => {
-            host_dmesg_indicates_oom(&String::from_utf8_lossy(&out.stdout), pid)
-        }
-        Ok(Ok(out)) => {
-            warn!(pid, exit_code = out.status.code(), "dmesg failed");
-            false
-        }
-        Ok(Err(e)) => {
-            warn!(pid, error = %e, "failed to run dmesg for OOM check");
-            false
-        }
-        Err(_) => {
-            warn!(pid, "host dmesg OOM check timed out");
-            false
-        }
-    }
-}
-
-/// Returns true if host dmesg output contains an OOM kill record for the
-/// given firecracker PID.  Checks that the character after the PID is not
-/// a digit to avoid prefix matches (e.g. pid=1234 must not match pid=12345).
-pub(super) fn host_dmesg_indicates_oom(dmesg: &str, pid: u32) -> bool {
-    if !dmesg.contains("oom-kill") {
-        return false;
-    }
-    let needle = format!("task=firecracker,pid={pid}");
-    let mut start = 0;
-    while let Some(pos) = dmesg[start..].find(&needle) {
-        let abs = start + pos + needle.len();
-        // Accept if needle is at end of string or next char is not a digit.
-        match dmesg.as_bytes().get(abs) {
-            Some(c) if c.is_ascii_digit() => {
-                // Prefix match (e.g. pid=1234 inside pid=12345) — keep searching.
-                start = abs;
-            }
-            _ => return true,
-        }
-    }
-    false
-}
-
-#[derive(Debug, thiserror::Error)]
-pub(super) enum StdoutDrainError {
-    #[error("failed to open host log file {path}: {source}")]
-    Open { path: PathBuf, source: io::Error },
-    #[error("failed to write stdout chunk to host log {path}: {source}")]
-    Write { path: PathBuf, source: io::Error },
-    #[error("failed to flush stdout log {path}: {source}")]
-    Flush { path: PathBuf, source: io::Error },
-}
-
-#[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
-pub(super) struct StdoutDrainReport {
-    pub(super) bytes_written: u64,
-    pub(super) chunk_truncated: bool,
-}
-
-/// Drain stdout chunks from the process receiver and write them to a host file.
-pub(super) async fn drain_stdout_to_file(
-    mut rx: ProcessOutputReceiver,
-    path: PathBuf,
-) -> Result<StdoutDrainReport, StdoutDrainError> {
-    let file = crate::log_file::open_append(&path, false).map(tokio::fs::File::from_std);
-    let mut file = match file {
-        Ok(f) => f,
-        Err(e) => {
-            return Err(StdoutDrainError::Open { path, source: e });
-        }
-    };
-    let mut report = StdoutDrainReport::default();
-    while let Some(chunk) = rx.recv().await {
-        report.bytes_written = report
-            .bytes_written
-            .saturating_add(u64::try_from(chunk.bytes.len()).unwrap_or(u64::MAX));
-        if chunk.truncated {
-            report.chunk_truncated = true;
-            warn!(path = %path.display(), "stdout stream chunk was truncated before host log write");
-        }
-        if let Err(e) = file.write_all(&chunk.bytes).await {
-            return Err(StdoutDrainError::Write { path, source: e });
-        }
-    }
-    // Flush to ensure the last blocking write completes before we return.
-    // tokio::fs::File::poll_write returns Ready before the blocking write finishes,
-    // so without flush the caller may observe incomplete file contents.
-    if let Err(e) = file.flush().await {
-        return Err(StdoutDrainError::Flush { path, source: e });
-    }
-    Ok(report)
-}
-
-pub(super) async fn append_stdout_stream_diagnostics_to_stream_log(
-    run_id: RunId,
-    path: &Path,
-    diagnostics: AgentStdoutStreamDiagnostics,
-) {
-    if diagnostics.is_empty() {
-        return;
-    }
-
-    if let Err(e) = append_stdout_stream_diagnostics(path, diagnostics).await {
-        warn!(
-            run_id = %run_id,
-            path = %path.display(),
-            error = %e,
-            "failed to append stdout stream diagnostic marker to host stream log"
-        );
-    }
-}
-
-pub(super) async fn append_stdout_stream_diagnostics(
-    path: &Path,
-    diagnostics: AgentStdoutStreamDiagnostics,
-) -> io::Result<()> {
-    if diagnostics.is_empty() {
-        return Ok(());
-    }
-
-    let mut file = tokio::fs::File::from_std(crate::log_file::open_append(path, true)?);
-
-    if file.metadata().await?.len() > 0 {
-        file.seek(SeekFrom::End(-1)).await?;
-        let mut last = [0u8; 1];
-        file.read_exact(&mut last).await?;
-        if last[0] != b'\n' {
-            file.write_all(b"\n").await?;
-        }
-    }
-    if diagnostics.chunk_truncated {
-        file.write_all(STDOUT_STREAM_LIMIT_MARKER).await?;
-    }
-    if diagnostics.stream_overflowed {
-        file.write_all(STDOUT_STREAM_OVERFLOW_MARKER).await?;
-    }
-    file.flush().await
-}
-
-/// Copy guest log files to host (best-effort, post-job).
-///
-/// The final system log copy keeps `system-*` as the guest-authored log. The
-/// supervised process stdout/stderr stream is written separately to
-/// `system-stream-*` in real time.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(super) enum GuestLogCopyFailureKind {
-    Failed,
-    SkippedAfterCancellation,
-}
-
-pub(super) fn guest_log_copy_failure_kind(cancelled: bool) -> GuestLogCopyFailureKind {
-    if cancelled {
-        GuestLogCopyFailureKind::SkippedAfterCancellation
-    } else {
-        GuestLogCopyFailureKind::Failed
-    }
-}
-
-pub(super) async fn copy_guest_logs(
-    sandbox: &dyn Sandbox,
-    context: &ExecutionContext,
-    log_paths: &LogPaths,
-    cancelled: bool,
-) {
-    let run_id = context.run_id;
-    let files = match [
-        guest_runtime_path(run_id, guest_contracts::runtime_paths::system_log_file)
-            .map(|path| (path, log_paths.system_log(run_id))),
-        guest_runtime_path(run_id, guest_contracts::runtime_paths::metrics_log_file)
-            .map(|path| (path, log_paths.metrics_log(run_id))),
-        guest_runtime_path(run_id, guest_contracts::runtime_paths::sandbox_ops_log_file)
-            .map(|path| (path, log_paths.sandbox_ops_log(run_id))),
-    ]
-    .into_iter()
-    .collect::<Result<Vec<_>, _>>()
-    {
-        Ok(files) => files,
-        Err(e) => {
-            warn!(run_id = %run_id, error = %e, "failed to resolve guest log paths");
-            return;
-        }
-    };
-
-    for (guest_path, host_path) in &files {
-        if let Err(e) = crate::log_file::validate_copy_destination(host_path) {
-            warn!(
-                run_id = %run_id,
-                error = %e,
-                guest_path = %guest_path,
-                host_path = %host_path.display(),
-                "skipping unsafe guest log destination"
-            );
-            continue;
-        }
-
-        if let Err(e) = sandbox
-            .copy_file(
-                guest_path,
-                host_path,
-                CopyFileOptions {
-                    max_bytes: GUEST_LOG_COPY_MAX_BYTES,
-                    timeout: DEFAULT_EXEC_TIMEOUT,
-                    missing_ok: true,
-                },
-            )
-            .await
-        {
-            match guest_log_copy_failure_kind(cancelled) {
-                GuestLogCopyFailureKind::SkippedAfterCancellation => {
-                    info!(run_id = %run_id, error = %e, guest_path = %guest_path, host_path = %host_path.display(), "guest log copy skipped after cancellation");
-                }
-                GuestLogCopyFailureKind::Failed => {
-                    warn!(run_id = %run_id, error = %e, guest_path = %guest_path, host_path = %host_path.display(), "failed to copy guest log");
-                }
-            }
-        }
-    }
-}
+#[cfg(test)]
+pub(super) use stdout_stream::{StdoutDrainError, append_stdout_stream_diagnostics};
