--- crates/runner/src/executor/tests/diagnostics_tests.rs
@@ -1,1035 +1,0 @@
-use std::collections::HashMap;
-use std::os::unix::fs::{PermissionsExt, symlink};
-use std::path::Path;
-use std::path::PathBuf;
-
-use guest_contracts::diagnostics::{
-    AgentFramework, CliObservedExitDiagnostic, CliObservedExitKind, CliTerminationDiagnostic,
-    CliTerminationReason, CliTerminationSignal, FAILURE_DIAGNOSTIC_SCHEMA_VERSION, FailureClass,
-    FailureDetailSource, FailureDiagnostic, FailureReason, PromptMetadata, SessionHistoryStatus,
-};
-use sandbox::{ExecTermination, ProcessExit, ProcessOutputChunk};
-use sandbox_mock::MockSandbox;
-
-use super::super::diagnostics::{
-    AgentStdoutStreamDiagnostics, GuestLogCopyFailureKind, StdoutDrainError,
-    append_stdout_stream_diagnostics, build_agent_env_diagnostics, build_agent_env_key_diagnostics,
-    copy_guest_logs, dmesg_indicates_oom, drain_stdout_to_file, guest_log_copy_failure_kind,
-    host_dmesg_indicates_oom, parse_agent_abnormal_exit_resource_diagnostics,
-    read_guest_cli_agent_session_id, read_guest_error_file, read_guest_failure_diagnostic_file,
-    should_collect_agent_abnormal_exit_diagnostics,
-    should_collect_unattributed_sigkill_resource_diagnostics,
-    should_log_agent_bootstrap_abnormal_exit_diagnostics,
-};
-use super::super::sandbox_run::post_job_cleanup;
-use super::super::{
-    AGENT_ENV_KEY_DIAGNOSTIC_LIMIT, AGENT_ENV_KEY_MAX_CHARS, GUEST_LOG_COPY_MAX_BYTES,
-    ResourceFailureKind, SMALL_GUEST_FILE_MAX_BYTES, STDOUT_STREAM_LIMIT_MARKER,
-    STDOUT_STREAM_OVERFLOW_MARKER, guest_runtime_path,
-};
-use super::support::{minimal_context, sandbox_exec_error, test_executor_config};
-use crate::ids::RunId;
-use crate::paths::LogPaths;
-
-fn mode(path: &Path) -> u32 {
-    std::fs::metadata(path).unwrap().permissions().mode() & 0o777
-}
-
-#[test]
-fn agent_env_diagnostics_sort_bounds_and_never_include_values() {
-    let mut bootstrap_env = HashMap::from([
-        ("BASH_ENV".to_string(), "super-secret-bash-env".to_string()),
-        ("NORMAL_KEY".to_string(), "normal-secret-value".to_string()),
-        ("VM0_RUN_ID".to_string(), "runner-secret-value".to_string()),
-        (
-            "VM0_SECRET_VALUES".to_string(),
-            "stored-secret-value".to_string(),
-        ),
-    ]);
-    let big_value = "largest-secret-value".repeat(20);
-    let big_value_len = big_value.len();
-    bootstrap_env.insert("BIG_VALUE".to_string(), big_value);
-    for index in 0..AGENT_ENV_KEY_DIAGNOSTIC_LIMIT {
-        bootstrap_env.insert(format!("ZZZ_{index:03}"), format!("value-{index}"));
-    }
-    bootstrap_env.insert(
-        format!("AAA_{}", "x".repeat(AGENT_ENV_KEY_MAX_CHARS * 4)),
-        "long-secret-value".to_string(),
-    );
-    let user_env = HashMap::from([("BASH_ENV".to_string(), "user-secret-bash-env".to_string())]);
-
-    let diagnostics = build_agent_env_diagnostics(&bootstrap_env, &user_env);
-
-    assert_eq!(diagnostics.env_count, AGENT_ENV_KEY_DIAGNOSTIC_LIMIT + 6);
-    assert!(diagnostics.env_bytes >= big_value_len);
-    assert_eq!(diagnostics.runner_owned_count, 2);
-    assert_eq!(
-        diagnostics.external_count,
-        AGENT_ENV_KEY_DIAGNOSTIC_LIMIT + 4
-    );
-    assert_eq!(diagnostics.suspicious_keys, vec!["BASH_ENV".to_string()]);
-    assert_eq!(diagnostics.largest_entries[0].key, "BIG_VALUE");
-    assert_eq!(diagnostics.largest_entries[0].value_len, big_value_len);
-    let env_pairs: Vec<(String, String)> = bootstrap_env.into_iter().collect();
-    let key_diagnostics = build_agent_env_key_diagnostics(&env_pairs);
-    assert_eq!(
-        key_diagnostics.logged_keys.len(),
-        AGENT_ENV_KEY_DIAGNOSTIC_LIMIT
-    );
-    assert_eq!(key_diagnostics.omitted_key_count, 6);
-    let mut sorted_logged_keys = key_diagnostics.logged_keys.clone();
-    sorted_logged_keys.sort();
-    assert_eq!(key_diagnostics.logged_keys, sorted_logged_keys);
-    let long_key = key_diagnostics
-        .logged_keys
-        .iter()
-        .find(|key| key.starts_with("AAA_"))
-        .expect("long key should be logged before the ZZZ keys");
-    assert_eq!(long_key.chars().count(), AGENT_ENV_KEY_MAX_CHARS + 3);
-    assert!(long_key.ends_with("..."));
-    let rendered = format!(
-        "{} {} {}",
-        diagnostics.suspicious_keys_csv(),
-        key_diagnostics.logged_keys_csv(),
-        diagnostics.largest_entries_csv()
-    );
-    assert!(rendered.contains("BASH_ENV"));
-    assert!(rendered.contains("BIG_VALUE"));
-    assert!(rendered.contains("VM0_RUN_ID"));
-    assert!(!rendered.contains("super-secret-bash-env"));
-    assert!(!rendered.contains("user-secret-bash-env"));
-    assert!(!rendered.contains("largest-secret-value"));
-    assert!(!rendered.contains("normal-secret-value"));
-    assert!(!rendered.contains("runner-secret-value"));
-    assert!(!rendered.contains("stored-secret-value"));
-    assert!(!rendered.contains("long-secret-value"));
-}
-
-#[test]
-fn bootstrap_abnormal_exit_diagnostics_allow_captured_stderr_without_resource_probe() {
-    let exit = ProcessExit::new(1, 126, Vec::new(), b"permission denied".to_vec());
-    let stderr = "permission denied";
-
-    assert!(should_log_agent_bootstrap_abnormal_exit_diagnostics(
-        false, &exit, None, None
-    ));
-    assert!(!should_collect_agent_abnormal_exit_diagnostics(
-        false, &exit, stderr, None, None
-    ));
-
-    let diagnostic = FailureDiagnostic::new(
-        FailureClass::CliNonzero,
-        AgentFramework::ClaudeCode,
-        PromptMetadata::from_prompt("/help"),
-    );
-    assert!(!should_log_agent_bootstrap_abnormal_exit_diagnostics(
-        false,
-        &exit,
-        Some(&diagnostic),
-        None
-    ));
-    assert!(!should_log_agent_bootstrap_abnormal_exit_diagnostics(
-        false,
-        &exit,
-        None,
-        Some("guest error")
-    ));
-
-    let successful_exit = ProcessExit::new(1, 0, Vec::new(), Vec::new());
-    assert!(!should_log_agent_bootstrap_abnormal_exit_diagnostics(
-        false,
-        &successful_exit,
-        None,
-        None
-    ));
-    let timed_out_exit = ProcessExit {
-        termination: ExecTermination::TimedOut,
-        ..ProcessExit::new(1, 0, Vec::new(), Vec::new())
-    };
-    assert!(!should_log_agent_bootstrap_abnormal_exit_diagnostics(
-        false,
-        &timed_out_exit,
-        None,
-        None
-    ));
-}
-
-fn fallback_cli_nonzero_diagnostic(exit_code: i32) -> FailureDiagnostic {
-    FailureDiagnostic::new(
-        FailureClass::CliNonzero,
-        AgentFramework::ClaudeCode,
-        PromptMetadata::from_prompt("/help"),
-    )
-    .with_cli_exit_code(exit_code)
-    .with_failure_detail_source(FailureDetailSource::FallbackExitCode)
-}
-
-#[test]
-fn unattributed_sigkill_resource_diagnostics_match_observed_sigkill() {
-    let exit = ProcessExit::new(1, 137, Vec::new(), Vec::new());
-    let diagnostic = fallback_cli_nonzero_diagnostic(137)
-        .with_cli_observed_exit(CliObservedExitDiagnostic::from_signal(libc::SIGKILL));
-
-    assert!(should_collect_unattributed_sigkill_resource_diagnostics(
-        false,
-        &exit,
-        Some(&diagnostic)
-    ));
-}
-
-#[test]
-fn unattributed_sigkill_resource_diagnostics_do_not_match_explicit_exit_137() {
-    let exit = ProcessExit::new(1, 137, Vec::new(), Vec::new());
-    let diagnostic = fallback_cli_nonzero_diagnostic(137)
-        .with_cli_observed_exit(CliObservedExitDiagnostic::from_exit_code(137));
-
-    assert!(!should_collect_unattributed_sigkill_resource_diagnostics(
-        false,
-        &exit,
-        Some(&diagnostic)
-    ));
-}
-
-#[test]
-fn unattributed_sigkill_resource_diagnostics_require_observed_sigkill() {
-    let exit = ProcessExit::new(1, 137, Vec::new(), Vec::new());
-    let diagnostic = fallback_cli_nonzero_diagnostic(137);
-
-    assert!(!should_collect_unattributed_sigkill_resource_diagnostics(
-        false,
-        &exit,
-        Some(&diagnostic)
-    ));
-}
-
-#[test]
-fn unattributed_sigkill_resource_diagnostics_require_fallback_detail_source() {
-    let exit = ProcessExit::new(1, 137, Vec::new(), Vec::new());
-    let diagnostic = FailureDiagnostic::new(
-        FailureClass::CliNonzero,
-        AgentFramework::ClaudeCode,
-        PromptMetadata::from_prompt("/help"),
-    )
-    .with_cli_exit_code(137)
-    .with_cli_observed_exit(CliObservedExitDiagnostic::from_signal(libc::SIGKILL));
-
-    assert!(!should_collect_unattributed_sigkill_resource_diagnostics(
-        false,
-        &exit,
-        Some(&diagnostic)
-    ));
-}
-
-#[test]
-fn unattributed_sigkill_resource_diagnostics_require_matching_exit_code() {
-    let exit = ProcessExit::new(1, 1, Vec::new(), Vec::new());
-    let diagnostic = fallback_cli_nonzero_diagnostic(137)
-        .with_cli_observed_exit(CliObservedExitDiagnostic::from_signal(libc::SIGKILL));
-
-    assert!(!should_collect_unattributed_sigkill_resource_diagnostics(
-        false,
-        &exit,
-        Some(&diagnostic)
-    ));
-}
-
-#[test]
-fn unattributed_sigkill_resource_diagnostics_require_standard_sigkill_mapping() {
-    let exit = ProcessExit::new(1, 99, Vec::new(), Vec::new());
-    let mut observed_exit = CliObservedExitDiagnostic::from_signal(libc::SIGKILL);
-    observed_exit.mapped_exit_code = 99;
-    let diagnostic = fallback_cli_nonzero_diagnostic(99).with_cli_observed_exit(observed_exit);
-
-    assert!(!should_collect_unattributed_sigkill_resource_diagnostics(
-        false,
-        &exit,
-        Some(&diagnostic)
-    ));
-}
-
-#[test]
-fn unattributed_sigkill_resource_diagnostics_require_signal_exit_shape() {
-    let exit = ProcessExit::new(1, 137, Vec::new(), Vec::new());
-    let diagnostic =
-        fallback_cli_nonzero_diagnostic(137).with_cli_observed_exit(CliObservedExitDiagnostic {
-            kind: CliObservedExitKind::Signal,
-            exit_code: Some(137),
-            signal_number: Some(libc::SIGKILL),
-            signal_name: Some("sigkill".to_string()),
-            mapped_exit_code: 137,
-        });
-
-    assert!(!should_collect_unattributed_sigkill_resource_diagnostics(
-        false,
-        &exit,
-        Some(&diagnostic)
-    ));
-}
-
-#[test]
-fn unattributed_sigkill_resource_diagnostics_require_unattributed_fallback_failure() {
-    let exit = ProcessExit::new(1, 137, Vec::new(), Vec::new());
-    let exit_with_process_diagnostic = ProcessExit {
-        diagnostic: "wait failed inside provider".to_string(),
-        ..ProcessExit::new(1, 137, Vec::new(), Vec::new())
-    };
-    let observed_sigkill_diagnostic = fallback_cli_nonzero_diagnostic(137)
-        .with_cli_observed_exit(CliObservedExitDiagnostic::from_signal(libc::SIGKILL));
-    let termination = CliTerminationDiagnostic::new(CliTerminationReason::StuckToolWatchdog)
-        .record_signal(CliTerminationSignal::Sigkill, Some(42), Some(1_000))
-        .with_observed_exit_code(137);
-    let attributed_diagnostic = observed_sigkill_diagnostic
-        .clone()
-        .with_cli_termination(termination);
-    let specific_diagnostic = FailureDiagnostic::new(
-        FailureClass::CliNonzero,
-        AgentFramework::ClaudeCode,
-        PromptMetadata::from_prompt("/help"),
-    )
-    .with_cli_exit_code(137)
-    .with_cli_observed_exit(CliObservedExitDiagnostic::from_signal(libc::SIGKILL))
-    .with_failure_detail_source(FailureDetailSource::ClaudeResult);
-    let non_cli_diagnostic = FailureDiagnostic::new(
-        FailureClass::CheckpointFailed,
-        AgentFramework::ClaudeCode,
-        PromptMetadata::from_prompt("/help"),
-    )
-    .with_cli_exit_code(137)
-    .with_cli_observed_exit(CliObservedExitDiagnostic::from_signal(libc::SIGKILL))
-    .with_failure_detail_source(FailureDetailSource::FallbackExitCode);
-    let timed_out_exit = ProcessExit {
-        termination: ExecTermination::TimedOut,
-        ..ProcessExit::new(1, 137, Vec::new(), Vec::new())
-    };
-
-    assert!(!should_collect_unattributed_sigkill_resource_diagnostics(
-        false,
-        &exit,
-        Some(&attributed_diagnostic)
-    ));
-    assert!(!should_collect_unattributed_sigkill_resource_diagnostics(
-        false,
-        &exit,
-        Some(&specific_diagnostic)
-    ));
-    assert!(!should_collect_unattributed_sigkill_resource_diagnostics(
-        false,
-        &exit,
-        Some(&non_cli_diagnostic)
-    ));
-    assert!(!should_collect_unattributed_sigkill_resource_diagnostics(
-        false,
-        &exit_with_process_diagnostic,
-        Some(&observed_sigkill_diagnostic)
-    ));
-    assert!(!should_collect_unattributed_sigkill_resource_diagnostics(
-        false,
-        &timed_out_exit,
-        Some(&observed_sigkill_diagnostic)
-    ));
-    assert!(!should_collect_unattributed_sigkill_resource_diagnostics(
-        true,
-        &exit,
-        Some(&observed_sigkill_diagnostic)
-    ));
-}
-
-#[test]
-fn dmesg_oom_positive() {
-    assert!(dmesg_indicates_oom(
-        "[  12.345] Out of memory: Killed process 1234 (claude)"
-    ));
-    assert!(dmesg_indicates_oom("oom-kill:constraint=CONSTRAINT_MEMCG"));
-    assert!(dmesg_indicates_oom("oom_reaper: reaped process 42"));
-}
-
-#[test]
-fn dmesg_oom_negative() {
-    assert!(!dmesg_indicates_oom(""));
-    // "Killed process" alone (without OOM context) should NOT match
-    assert!(!dmesg_indicates_oom("Killed process 42 (node)"));
-    assert!(!dmesg_indicates_oom("normal kernel log output"));
-    assert!(!dmesg_indicates_oom("[  1.000] eth0: link up"));
-    assert!(!dmesg_indicates_oom("task killed by signal 15"));
-    // substring "oom" in unrelated words should not match
-    assert!(!dmesg_indicates_oom("the room is full"));
-}
-
-#[test]
-fn dmesg_oom_case_insensitive() {
-    assert!(dmesg_indicates_oom("Out Of Memory: killed process 99"));
-    assert!(!dmesg_indicates_oom("Killed process 99 (agent)"));
-    assert!(dmesg_indicates_oom("OOM-kill: constraint=MEMCG"));
-}
-
-#[test]
-fn abnormal_exit_resource_diagnostics_classifies_production_rootfs_full_sample() {
-    let diagnostics = parse_agent_abnormal_exit_resource_diagnostics(
-        r#"
-/dev/root       7.8G  7.4G   20K 100% /
-/dev/vdb         16G   24K   15G   1% /home/user/workspace
-Mem:            3934        3310         255           0         552         624
-Swap:              0           0           0
-"#,
-    )
-    .expect("resource diagnostics should parse production sample");
-
-    assert_eq!(
-        diagnostics.failure_kind,
-        Some(ResourceFailureKind::GuestRootFilesystemFull)
-    );
-    assert_eq!(diagnostics.guest_root_fs_used_percent, Some(100));
-    assert_eq!(diagnostics.guest_root_fs_available_kb, Some(20));
-    assert_eq!(diagnostics.guest_workspace_fs_used_percent, Some(1));
-    assert_eq!(diagnostics.guest_memory_available_mb, Some(624));
-}
-
-#[test]
-fn abnormal_exit_resource_diagnostics_does_not_classify_workspace_full_as_rootfs_full() {
-    let diagnostics = parse_agent_abnormal_exit_resource_diagnostics(
-        r#"
-/dev/root       16G  4.0G   12G  25% /
-/dev/vdb        16G   16G     0 100% /home/user/workspace
-Mem:          4096  1024  2048  0 1024  3072
-"#,
-    )
-    .expect("resource diagnostics should parse workspace sample");
-
-    assert_eq!(diagnostics.failure_kind, None);
-    assert_eq!(diagnostics.guest_root_fs_used_percent, Some(25));
-    assert_eq!(diagnostics.guest_workspace_fs_used_percent, Some(100));
-    assert_eq!(diagnostics.guest_memory_available_mb, Some(3072));
-}
-
-#[test]
-fn abnormal_exit_resource_diagnostics_does_not_classify_normal_parseable_df_output() {
-    let diagnostics = parse_agent_abnormal_exit_resource_diagnostics(
-        r#"
-Filesystem     1024-blocks    Used Available Capacity Mounted on
-/dev/root         16447660 4194304  12253356      26% /
-/dev/vdb          16447660      24  16447636       1% /home/user/workspace
-MemAvailable:    1048576 kB
-"#,
-    )
-    .expect("resource diagnostics should parse parseable df output");
-
-    assert_eq!(diagnostics.failure_kind, None);
-    assert_eq!(diagnostics.guest_root_fs_used_percent, Some(26));
-    assert_eq!(diagnostics.guest_root_fs_available_kb, Some(12_253_356));
-    assert_eq!(diagnostics.guest_workspace_fs_used_percent, Some(1));
-    assert_eq!(diagnostics.guest_memory_available_mb, Some(1024));
-}
-
-#[test]
-fn abnormal_exit_resource_diagnostics_ignores_malformed_output() {
-    let diagnostics = parse_agent_abnormal_exit_resource_diagnostics(
-        r#"
-df: unavailable
-Mem: not numeric
-/dev/root missing columns
-"#,
-    );
-
-    assert_eq!(diagnostics, None);
-}
-
-#[test]
-fn abnormal_exit_resource_diagnostics_ignores_invalid_available_values() {
-    let diagnostics = parse_agent_abnormal_exit_resource_diagnostics(
-        r#"
-/dev/root       16G  4.0G   NaNK  25% /
-/dev/vdb        16G   24K    -1K   1% /home/user/workspace
-"#,
-    )
-    .expect("resource diagnostics should keep valid percentages");
-
-    assert_eq!(diagnostics.failure_kind, None);
-    assert_eq!(diagnostics.guest_root_fs_used_percent, Some(25));
-    assert_eq!(diagnostics.guest_root_fs_available_kb, None);
-    assert_eq!(diagnostics.guest_workspace_fs_used_percent, Some(1));
-}
-
-/// Real `sudo dmesg | grep 'oom-kill'` output captured from prod-3.
-const PROD3_OOM_GREP: &str = "\
-        [1718300.650867] fc_vcpu 0 invoked oom-killer: gfp_mask=0xcc0(GFP_KERNEL), order=0, oom_score_adj=0\n\
-        [1718300.651117] oom-kill:constraint=CONSTRAINT_MEMCG,nodemask=(null),cpuset=vm0-runner-v0.45.6.service,mems_allowed=0,oom_memcg=/system.slice/vm0-runner-v0.45.6.service,task_memcg=/system.slice/vm0-runner-v0.45.6.service,task=firecracker,pid=586629,uid=1000";
-
-#[test]
-fn host_oom_matches_real_prod3_output() {
-    assert!(host_dmesg_indicates_oom(PROD3_OOM_GREP, 586629));
-}
-
-#[test]
-fn host_oom_no_match_different_pid() {
-    assert!(!host_dmesg_indicates_oom(PROD3_OOM_GREP, 12345));
-}
-
-#[test]
-fn host_oom_no_match_different_process() {
-    // Same structure as prod-3 but task=node instead of task=firecracker
-    let dmesg = "[1718300.651117] oom-kill:constraint=CONSTRAINT_MEMCG,\
-            task=node,pid=586629,uid=1000";
-    assert!(!host_dmesg_indicates_oom(dmesg, 586629));
-}
-
-#[test]
-fn host_oom_no_match_empty() {
-    assert!(!host_dmesg_indicates_oom("", 12345));
-}
-
-#[test]
-fn host_oom_no_match_without_oom_kill() {
-    // Has the PID pattern but no oom-kill keyword
-    let dmesg = "[1718300.651117] task=firecracker,pid=12345,uid=1000 started";
-    assert!(!host_dmesg_indicates_oom(dmesg, 12345));
-}
-
-#[test]
-fn host_oom_no_prefix_match() {
-    // pid=58662 must NOT match pid=586629
-    assert!(!host_dmesg_indicates_oom(PROD3_OOM_GREP, 58662));
-}
-
-#[test]
-fn host_oom_pid_at_end_of_line() {
-    // PID at end of string (no trailing comma) — edge case
-    let dmesg = "[0.0] oom-kill:constraint=CONSTRAINT_MEMCG,task=firecracker,pid=42";
-    assert!(host_dmesg_indicates_oom(dmesg, 42));
-    assert!(!host_dmesg_indicates_oom(dmesg, 4));
-}
-
-#[tokio::test]
-async fn read_guest_error_file_returns_content() {
-    let sandbox = MockSandbox::new("test");
-    sandbox.push_read_file_result(Ok(Some(b"checkpoint error: disk full".to_vec())));
-    let msg = read_guest_error_file(&sandbox, RunId::nil()).await;
-    assert_eq!(msg.as_deref(), Some("checkpoint error: disk full"));
-    let calls = sandbox.read_file_calls();
-    assert_eq!(calls.len(), 1);
-    assert_eq!(
-        calls[0].path,
-        guest_runtime_path(
-            RunId::nil(),
-            guest_contracts::runtime_paths::checkpoint_error_file
-        )
-        .unwrap()
-    );
-    assert_eq!(calls[0].max_bytes, SMALL_GUEST_FILE_MAX_BYTES);
-}
-
-#[tokio::test]
-async fn read_guest_error_file_returns_none_on_missing_file() {
-    let sandbox = MockSandbox::new("test");
-    sandbox.push_read_file_result(Ok(None));
-    let msg = read_guest_error_file(&sandbox, RunId::nil()).await;
-    assert!(msg.is_none());
-}
-
-#[tokio::test]
-async fn read_guest_error_file_returns_none_on_empty_content() {
-    let sandbox = MockSandbox::new("test");
-    sandbox.push_read_file_result(Ok(Some(b"   \n  ".to_vec())));
-    let msg = read_guest_error_file(&sandbox, RunId::nil()).await;
-    assert!(msg.is_none());
-}
-
-#[tokio::test]
-async fn read_guest_error_file_returns_none_on_exec_error() {
-    let sandbox = MockSandbox::new("test");
-    sandbox.push_read_file_result(Err(sandbox_exec_error("vsock timeout")));
-    let msg = read_guest_error_file(&sandbox, RunId::nil()).await;
-    assert!(msg.is_none());
-}
-
-#[tokio::test]
-async fn read_guest_cli_agent_session_id_returns_trimmed_content_from_runtime_path() {
-    let sandbox = MockSandbox::new("test");
-    sandbox.push_read_file_result(Ok(Some(b" session-abc \n".to_vec())));
-
-    let session_id = read_guest_cli_agent_session_id(&sandbox, RunId::nil()).await;
-
-    assert_eq!(session_id.as_deref(), Some("session-abc"));
-    let calls = sandbox.read_file_calls();
-    assert_eq!(calls.len(), 1);
-    assert_eq!(
-        calls[0].path,
-        guest_runtime_path(
-            RunId::nil(),
-            guest_contracts::runtime_paths::session_id_file
-        )
-        .unwrap()
-    );
-    assert_eq!(calls[0].max_bytes, SMALL_GUEST_FILE_MAX_BYTES);
-}
-
-#[tokio::test]
-async fn read_guest_cli_agent_session_id_returns_none_on_missing_or_empty_file() {
-    let missing = MockSandbox::new("test");
-    missing.push_read_file_result(Ok(None));
-    assert!(
-        read_guest_cli_agent_session_id(&missing, RunId::nil())
-            .await
-            .is_none()
-    );
-
-    let empty = MockSandbox::new("test");
-    empty.push_read_file_result(Ok(Some(b" \n ".to_vec())));
-    assert!(
-        read_guest_cli_agent_session_id(&empty, RunId::nil())
-            .await
-            .is_none()
-    );
-}
-
-#[tokio::test]
-async fn read_guest_cli_agent_session_id_rejects_invalid_content() {
-    let sandbox = MockSandbox::new("test");
-    sandbox.push_read_file_result(Ok(Some(b" ../session \n".to_vec())));
-
-    let session_id = read_guest_cli_agent_session_id(&sandbox, RunId::nil()).await;
-
-    assert!(session_id.is_none());
-}
-
-#[tokio::test]
-async fn read_guest_cli_agent_session_id_rejects_overlong_content() {
-    let sandbox = MockSandbox::new("test");
-    sandbox.push_read_file_result(Ok(Some("a".repeat(129).into_bytes())));
-
-    let session_id = read_guest_cli_agent_session_id(&sandbox, RunId::nil()).await;
-
-    assert!(session_id.is_none());
-}
-
-#[tokio::test]
-async fn read_guest_failure_diagnostic_file_returns_valid_diagnostic() {
-    let sandbox = MockSandbox::new("test");
-    let diagnostic = FailureDiagnostic::new(
-        FailureClass::CliNonzero,
-        AgentFramework::ClaudeCode,
-        PromptMetadata::from_prompt("/help"),
-    )
-    .with_cli_exit_code(1)
-    .with_failure_detail_source(FailureDetailSource::ClaudeResult)
-    .with_failure_reason(FailureReason::ProviderOverloaded)
-    .with_session_history_status(SessionHistoryStatus::Present);
-    sandbox.push_read_file_result(Ok(Some(serde_json::to_vec(&diagnostic).unwrap())));
-
-    let read = read_guest_failure_diagnostic_file(&sandbox, RunId::nil()).await;
-
-    assert_eq!(read, Some(diagnostic));
-    let calls = sandbox.read_file_calls();
-    assert_eq!(calls.len(), 1);
-    assert_eq!(
-        calls[0].path,
-        guest_runtime_path(
-            RunId::nil(),
-            guest_contracts::runtime_paths::failure_diagnostic_file
-        )
-        .unwrap()
-    );
-    assert_eq!(calls[0].max_bytes, SMALL_GUEST_FILE_MAX_BYTES);
-}
-
-#[tokio::test]
-async fn read_guest_failure_diagnostic_file_returns_none_on_missing_file() {
-    let sandbox = MockSandbox::new("test");
-    sandbox.push_read_file_result(Ok(None));
-
-    let diagnostic = read_guest_failure_diagnostic_file(&sandbox, RunId::nil()).await;
-
-    assert!(diagnostic.is_none());
-}
-
-#[tokio::test]
-async fn read_guest_failure_diagnostic_file_returns_none_on_empty_content() {
-    let sandbox = MockSandbox::new("test");
-    sandbox.push_read_file_result(Ok(Some(b" \n\t".to_vec())));
-
-    let diagnostic = read_guest_failure_diagnostic_file(&sandbox, RunId::nil()).await;
-
-    assert!(diagnostic.is_none());
-}
-
-#[tokio::test]
-async fn read_guest_failure_diagnostic_file_returns_none_on_malformed_json() {
-    let sandbox = MockSandbox::new("test");
-    sandbox.push_read_file_result(Ok(Some(b"{not-json".to_vec())));
-
-    let diagnostic = read_guest_failure_diagnostic_file(&sandbox, RunId::nil()).await;
-
-    assert!(diagnostic.is_none());
-}
-
-#[tokio::test]
-async fn read_guest_failure_diagnostic_file_returns_none_on_unsupported_schema() {
-    let sandbox = MockSandbox::new("test");
-    let mut diagnostic = FailureDiagnostic::new(
-        FailureClass::CliNonzero,
-        AgentFramework::ClaudeCode,
-        PromptMetadata::from_prompt("/help"),
-    );
-    diagnostic.schema_version = FAILURE_DIAGNOSTIC_SCHEMA_VERSION + 1;
-    sandbox.push_read_file_result(Ok(Some(serde_json::to_vec(&diagnostic).unwrap())));
-
-    let diagnostic = read_guest_failure_diagnostic_file(&sandbox, RunId::nil()).await;
-
-    assert!(diagnostic.is_none());
-}
-
-#[tokio::test]
-async fn read_guest_failure_diagnostic_file_returns_none_on_read_error() {
-    let sandbox = MockSandbox::new("test");
-    sandbox.push_read_file_result(Err(sandbox_exec_error("vsock timeout")));
-
-    let diagnostic = read_guest_failure_diagnostic_file(&sandbox, RunId::nil()).await;
-
-    assert!(diagnostic.is_none());
-}
-
-#[tokio::test]
-async fn read_guest_failure_diagnostic_file_returns_none_on_oversized_content() {
-    let sandbox = MockSandbox::new("test");
-    sandbox.push_read_file_result(Ok(Some(vec![
-        b' ';
-        SMALL_GUEST_FILE_MAX_BYTES as usize + 1
-    ])));
-
-    let diagnostic = read_guest_failure_diagnostic_file(&sandbox, RunId::nil()).await;
-
-    assert!(diagnostic.is_none());
-}
-
-// -----------------------------------------------------------------------
-// copy_guest_logs tests
-// -----------------------------------------------------------------------
-
-#[test]
-fn guest_log_copy_failure_kind_tracks_cancellation() {
-    assert_eq!(
-        guest_log_copy_failure_kind(false),
-        GuestLogCopyFailureKind::Failed
-    );
-    assert_eq!(
-        guest_log_copy_failure_kind(true),
-        GuestLogCopyFailureKind::SkippedAfterCancellation
-    );
-}
-
-#[tokio::test]
-async fn copy_guest_logs_writes_files_to_host() {
-    let dir = tempfile::tempdir().unwrap();
-    let log_paths = LogPaths::new(dir.path().to_path_buf());
-    let sandbox = MockSandbox::new("test");
-    let ctx = minimal_context();
-
-    let system_stream_log_path = log_paths.system_stream_log(ctx.run_id);
-    tokio::fs::write(&system_stream_log_path, b"transient host-streamed stdout\n")
-        .await
-        .unwrap();
-
-    // Queue guest-copy results: system log + metrics log + sandbox ops log.
-    sandbox.push_copy_file_result(Ok(b"system log line 1\nsystem log line 2\n".to_vec()));
-    sandbox.push_copy_file_result(Ok(b"{\"cpu\":0.5}\n".to_vec()));
-    sandbox.push_copy_file_result(Ok(
-        b"{\"action_type\":\"final_telemetry_upload\",\"duration_ms\":10,\"success\":true}\n"
-            .to_vec(),
-    ));
-
-    copy_guest_logs(&sandbox, &ctx, &log_paths, false).await;
-
-    let system_log = tokio::fs::read_to_string(log_paths.system_log(ctx.run_id))
-        .await
-        .unwrap();
-    assert_eq!(system_log, "system log line 1\nsystem log line 2\n");
-    let system_stream_log = tokio::fs::read_to_string(system_stream_log_path)
-        .await
-        .unwrap();
-    assert_eq!(system_stream_log, "transient host-streamed stdout\n");
-
-    let metrics_log = tokio::fs::read_to_string(log_paths.metrics_log(ctx.run_id))
-        .await
-        .unwrap();
-    assert_eq!(metrics_log, "{\"cpu\":0.5}\n");
-
-    let sandbox_ops_log = tokio::fs::read_to_string(log_paths.sandbox_ops_log(ctx.run_id))
-        .await
-        .unwrap();
-    assert!(sandbox_ops_log.contains("final_telemetry_upload"));
-
-    let calls = sandbox.copy_file_calls();
-    assert_eq!(calls.len(), 3);
-    assert_eq!(
-        calls[2].path,
-        guest_runtime_path(
-            ctx.run_id,
-            guest_contracts::runtime_paths::sandbox_ops_log_file
-        )
-        .unwrap()
-    );
-    assert_eq!(calls[2].host_path, log_paths.sandbox_ops_log(ctx.run_id));
-    assert_eq!(calls[0].max_bytes, GUEST_LOG_COPY_MAX_BYTES);
-    assert_eq!(calls[1].max_bytes, GUEST_LOG_COPY_MAX_BYTES);
-    assert_eq!(calls[2].max_bytes, GUEST_LOG_COPY_MAX_BYTES);
-}
-
-#[tokio::test]
-async fn copy_guest_logs_skips_unsafe_destination_and_continues_other_logs() {
-    let dir = tempfile::tempdir().unwrap();
-    let log_paths = LogPaths::new(dir.path().to_path_buf());
-    let sandbox = MockSandbox::new("test");
-    let ctx = minimal_context();
-    let unsafe_target = dir.path().join("unsafe-target.log");
-    symlink(&unsafe_target, log_paths.system_log(ctx.run_id)).unwrap();
-
-    sandbox.push_copy_file_result(Ok(b"{\"cpu\":0.5}\n".to_vec()));
-    sandbox.push_copy_file_result(Ok(b"{\"action_type\":\"cleanup\"}\n".to_vec()));
-
-    copy_guest_logs(&sandbox, &ctx, &log_paths, false).await;
-
-    assert!(!unsafe_target.exists());
-    assert_eq!(
-        tokio::fs::read_to_string(log_paths.metrics_log(ctx.run_id))
-            .await
-            .unwrap(),
-        "{\"cpu\":0.5}\n"
-    );
-    assert_eq!(
-        tokio::fs::read_to_string(log_paths.sandbox_ops_log(ctx.run_id))
-            .await
-            .unwrap(),
-        "{\"action_type\":\"cleanup\"}\n"
-    );
-
-    let calls = sandbox.copy_file_calls();
-    assert_eq!(calls.len(), 2);
-    assert_eq!(calls[0].host_path, log_paths.metrics_log(ctx.run_id));
-    assert_eq!(calls[1].host_path, log_paths.sandbox_ops_log(ctx.run_id));
-}
-
-#[tokio::test]
-async fn copy_guest_logs_keeps_existing_logs_when_sandbox_ops_missing() {
-    let dir = tempfile::tempdir().unwrap();
-    let log_paths = LogPaths::new(dir.path().to_path_buf());
-    let sandbox = MockSandbox::new("test");
-    let ctx = minimal_context();
-
-    sandbox.push_copy_file_result(Ok(b"system log\n".to_vec()));
-    sandbox.push_copy_file_result(Ok(b"{\"cpu\":0.5}\n".to_vec()));
-
-    copy_guest_logs(&sandbox, &ctx, &log_paths, false).await;
-
-    let system_log = tokio::fs::read_to_string(log_paths.system_log(ctx.run_id))
-        .await
-        .unwrap();
-    assert_eq!(system_log, "system log\n");
-
-    let metrics_log = tokio::fs::read_to_string(log_paths.metrics_log(ctx.run_id))
-        .await
-        .unwrap();
-    assert_eq!(metrics_log, "{\"cpu\":0.5}\n");
-    assert!(!log_paths.sandbox_ops_log(ctx.run_id).exists());
-
-    let calls = sandbox.copy_file_calls();
-    assert_eq!(calls.len(), 3);
-    assert!(
-        calls[2].missing_ok,
-        "missing sandbox ops log should be a best-effort no-op"
-    );
-}
-
-#[tokio::test]
-async fn copy_guest_logs_skips_on_nonzero_exit() {
-    let dir = tempfile::tempdir().unwrap();
-    let log_paths = LogPaths::new(dir.path().to_path_buf());
-    let sandbox = MockSandbox::new("test");
-    let ctx = minimal_context();
-
-    // Copy fails (file doesn't exist in guest).
-    sandbox.push_copy_file_result(Err(sandbox_exec_error("No such file")));
-    sandbox.push_copy_file_result(Err(sandbox_exec_error("No such file")));
-    sandbox.push_copy_file_result(Err(sandbox_exec_error("No such file")));
-
-    copy_guest_logs(&sandbox, &ctx, &log_paths, false).await;
-
-    // Host files should not be created
-    assert!(!log_paths.system_log(ctx.run_id).exists());
-    assert!(!log_paths.metrics_log(ctx.run_id).exists());
-    assert!(!log_paths.sandbox_ops_log(ctx.run_id).exists());
-}
-
-#[tokio::test]
-async fn copy_guest_logs_skips_on_exec_error() {
-    let dir = tempfile::tempdir().unwrap();
-    let log_paths = LogPaths::new(dir.path().to_path_buf());
-    let sandbox = MockSandbox::new("test");
-    let ctx = minimal_context();
-
-    sandbox.push_copy_file_result(Err(sandbox_exec_error("vsock down")));
-    sandbox.push_copy_file_result(Err(sandbox_exec_error("vsock down")));
-    sandbox.push_copy_file_result(Err(sandbox_exec_error("vsock down")));
-
-    copy_guest_logs(&sandbox, &ctx, &log_paths, false).await;
-
-    assert!(!log_paths.system_log(ctx.run_id).exists());
-    assert!(!log_paths.metrics_log(ctx.run_id).exists());
-    assert!(!log_paths.sandbox_ops_log(ctx.run_id).exists());
-}
-
-#[tokio::test]
-async fn post_job_cleanup_appends_stream_markers_after_guest_log_copy() {
-    let dir = tempfile::tempdir().unwrap();
-    let config = test_executor_config(dir.path()).await;
-    let sandbox = MockSandbox::new("test");
-    let ctx = minimal_context();
-    let system_log_path = config.log_paths.system_log(ctx.run_id);
-    let system_stream_log_path = config.log_paths.system_stream_log(ctx.run_id);
-
-    tokio::fs::write(&system_stream_log_path, b"transient host-streamed stdout\n")
-        .await
-        .unwrap();
-    sandbox.push_copy_file_result(Ok(b"guest system log".to_vec()));
-
-    post_job_cleanup(
-        &sandbox,
-        &config,
-        &ctx,
-        "10.0.0.1",
-        false,
-        AgentStdoutStreamDiagnostics {
-            bytes_written: 0,
-            chunk_truncated: true,
-            stream_overflowed: true,
-        },
-    )
-    .await
-    .unwrap();
-
-    let system_log = tokio::fs::read(&system_log_path).await.unwrap();
-    assert_eq!(system_log, b"guest system log");
-    let system_stream_log = tokio::fs::read(&system_stream_log_path).await.unwrap();
-    let mut expected_stream_log = b"transient host-streamed stdout\n".to_vec();
-    expected_stream_log.extend_from_slice(STDOUT_STREAM_LIMIT_MARKER);
-    expected_stream_log.extend_from_slice(STDOUT_STREAM_OVERFLOW_MARKER);
-    assert_eq!(system_stream_log, expected_stream_log);
-}
-
-// -----------------------------------------------------------------------
-// drain_stdout_to_file tests
-// -----------------------------------------------------------------------
-
-#[tokio::test]
-async fn drain_stdout_writes_chunks_to_file() {
-    let dir = tempfile::tempdir().unwrap();
-    let path = dir.path().join("stdout.log");
-
-    let (tx, rx) = tokio::sync::mpsc::channel(2);
-    tx.send(ProcessOutputChunk {
-        bytes: b"chunk 1\n".to_vec(),
-        truncated: false,
-    })
-    .await
-    .unwrap();
-    tx.send(ProcessOutputChunk {
-        bytes: b"chunk 2\n".to_vec(),
-        truncated: false,
-    })
-    .await
-    .unwrap();
-    drop(tx); // close channel
-
-    let report = drain_stdout_to_file(rx, path.clone()).await.unwrap();
-
-    let content = tokio::fs::read_to_string(&path).await.unwrap();
-    assert_eq!(content, "chunk 1\nchunk 2\n");
-    assert_eq!(report.bytes_written, 16);
-    assert!(!report.chunk_truncated);
-    assert_eq!(mode(&path), 0o600);
-}
-
-#[tokio::test]
-async fn drain_stdout_reports_truncated_chunk_without_changing_bytes() {
-    let dir = tempfile::tempdir().unwrap();
-    let path = dir.path().join("stdout.log");
-
-    let (tx, rx) = tokio::sync::mpsc::channel(1);
-    tx.send(ProcessOutputChunk {
-        bytes: b"partial chunk".to_vec(),
-        truncated: true,
-    })
-    .await
-    .unwrap();
-    drop(tx);
-
-    let report = drain_stdout_to_file(rx, path.clone()).await.unwrap();
-
-    let content = tokio::fs::read(&path).await.unwrap();
-    assert_eq!(content, b"partial chunk");
-    assert_eq!(report.bytes_written, 13);
-    assert!(report.chunk_truncated);
-}
-
-#[tokio::test]
-async fn drain_stdout_empty_channel() {
-    let dir = tempfile::tempdir().unwrap();
-    let path = dir.path().join("empty.log");
-
-    let (_tx, rx) = tokio::sync::mpsc::channel::<ProcessOutputChunk>(1);
-    drop(_tx);
-
-    let report = drain_stdout_to_file(rx, path.clone()).await.unwrap();
-
-    let content = tokio::fs::read_to_string(&path).await.unwrap();
-    assert!(content.is_empty());
-    assert_eq!(report.bytes_written, 0);
-    assert!(!report.chunk_truncated);
-}
-
-#[tokio::test]
-async fn drain_stdout_invalid_path_returns_error() {
-    let (_tx, rx) = tokio::sync::mpsc::channel::<ProcessOutputChunk>(1);
-    drop(_tx);
-    let error = drain_stdout_to_file(rx, PathBuf::from("/dev/null/impossible/file"))
-        .await
-        .unwrap_err();
-    assert!(matches!(error, StdoutDrainError::Open { .. }));
-}
-
-#[tokio::test]
-async fn append_stdout_stream_diagnostics_noops_when_empty() {
-    let dir = tempfile::tempdir().unwrap();
-    let path = dir.path().join("stdout.log");
-
-    append_stdout_stream_diagnostics(&path, AgentStdoutStreamDiagnostics::default())
-        .await
-        .unwrap();
-
-    assert!(!path.exists());
-}
-
-#[tokio::test]
-async fn append_stdout_stream_diagnostics_writes_markers() {
-    let dir = tempfile::tempdir().unwrap();
-    let path = dir.path().join("stdout.log");
-    tokio::fs::write(&path, b"guest system log without newline")
-        .await
-        .unwrap();
-
-    append_stdout_stream_diagnostics(
-        &path,
-        AgentStdoutStreamDiagnostics {
-            bytes_written: 0,
-            chunk_truncated: true,
-            stream_overflowed: true,
-        },
-    )
-    .await
-    .unwrap();
-
-    let content = tokio::fs::read(&path).await.unwrap();
-    let mut expected = b"guest system log without newline\n".to_vec();
-    expected.extend_from_slice(STDOUT_STREAM_LIMIT_MARKER);
-    expected.extend_from_slice(STDOUT_STREAM_OVERFLOW_MARKER);
-    assert_eq!(content, expected);
-    assert_eq!(mode(&path), 0o600);
-}
