+++ crates/runner/src/executor/tests/diagnostics_tests/mod.rs
@@ -1,0 +1,7 @@
+mod environment;
+mod exit;
+mod guest_files;
+mod guest_logs;
+mod oom;
+mod resource;
+mod stdout_stream;
